+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Version history
 
+- 0.3.4 – (2026/04/06) Remove all domain logic from openmsxControl;
 - 0.3.3 – (2026/04/02) Step Over;
 - 0.3.2 – (2026/04/01) Step Out;
-- 0.3.1 – (2026/03/30) Basic debugging and;
+- 0.3.1 – (2026/03/30) Basic debugging and Step In;
 - 0.3.0 – (2026/03/27) Stack tracing support;
 - 0.2.0 – (2026/03/26) Logger implementation and array support;
 - 0.1.0 – (2026/03/25) Initial release.